producer/requirements.txt
pika
fastapi
retrying
uvicorn
pythonping
mysql-connector-python